+++ src/main/java/vn/flast/repositories/TransporterRepository.java
@@ -3,5 +3,5 @@ package vn.flast.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.flast.models.Transporter;
 
-public interface TransporterRepository extends JpaRepository<Transporter, Long> {
+public interface TransporterRepository extends JpaRepository<Transporter, Integer> {
 }
